@@ -8,6 +8,7 @@ using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Integration.AspNet.Core;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace LivePersonAgentBot
 {
@@ -23,8 +24,11 @@ namespace LivePersonAgentBot
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
 
+            services.AddMvc(option => option.EnableEndpointRouting = false);
+
+            // MemoryStorage is used for testing purposes only. You must use durable storage in production implementation
             var storage = new MemoryStorage();
             // Create the User state passing in the storage layer.
             var userState = new UserState(storage);
@@ -44,7 +48,7 @@ namespace LivePersonAgentBot
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
