@@ -1,20 +1,17 @@
-﻿using System;
-using Microsoft.Bot.Builder;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Bot.Builder;
 using Microsoft.Bot.Schema;
-using Microsoft.Extensions.Configuration;
-using System.Linq;
-using LivePersonAgentBot.Bots;
-using System.Net.Http;
-using Newtonsoft.Json;
-using System.Text;
-using System.Collections.Concurrent;
 using Newtonsoft.Json.Linq;
 
 namespace LivePersonAgentBot
 {
-    public class HandoffMiddleware : Microsoft.Bot.Builder.IMiddleware
+    public class HandoffMiddleware : IMiddleware
     {
         public HandoffMiddleware()
         {
@@ -24,7 +21,8 @@ namespace LivePersonAgentBot
         {
             turnContext.OnSendActivities(async (sendTurnContext, activities, nextSend) =>
             {
-                Activity handOffEventActivity = activities.Where((activity) => activity.Type == ActivityTypes.Event && activity.Name == HandoffEventNames.InitiateHandoff).FirstOrDefault();
+                Activity handOffEventActivity = activities.Where((activity) => activity.Type == ActivityTypes.Event &&
+                                                                               activity.Name == HandoffEventNames.InitiateHandoff).FirstOrDefault();
 
                 if (handOffEventActivity != null)
                 {
@@ -35,20 +33,7 @@ namespace LivePersonAgentBot
                     {
                         skill = "default";
                     }
-#if false // using string literal - works but more verbose
-                    var channelData = @"
-                    {
-                    ""action"": {
-                        ""name"": ""TRANSFER"",
-                        ""parameters"": {
-                            ""skill"": ""credit cards""
-                        }
-                    }
-                    }";
-                    var activity = new Activity("message");
-                    activity.Text = "";
-                    activity.ChannelData = JObject.Parse(channelData);
-#else // nicer
+
                     var channelData = new { action =
                         new {
                             name = "TRANSFER",
@@ -59,7 +44,6 @@ namespace LivePersonAgentBot
                     var activity = new Activity("message");
                     activity.Text = "";
                     activity.ChannelData = channelData;
-#endif
 
                     await sendTurnContext.SendActivityAsync(activity).ConfigureAwait(false);
                 }
